+++ R/mf_pkg_utils.R
@@ -8,10 +8,7 @@
 #' @importFrom stats aggregate median na.omit quantile runif sd
 
 
-# Rcpp stuff
-#' @useDynLib mapsf,.registration = TRUE
-#' @importFrom Rcpp evalCpp
-NULL
+
 
 #' @importFrom utils globalVariables
 .gmapsf <- new.env(parent = emptyenv())
@@ -203,4 +200,28 @@ interleg <- function(txt = c("legend", "Legend")) {
       call. = FALSE
     )
   }
+}
+
+
+# shadow around the labels
+#' @name shadowtext
+#' @title shadowtext
+#' @description shadowtext
+#' @param x lon
+#' @param y lat
+#' @param labels labels
+#' @param col col
+#' @param bg bg
+#' @param theta number of iteration
+#' @param r radius
+#' @param ... other txt params
+#' @noRd
+shadowtext <- function(x, y = NULL, labels, col = "white", bg = "black",
+                       theta = seq(0, 2 * pi, length.out = 50), r = 0.1, ...) {
+  xo <- r * strwidth("A")
+  yo <- r * strheight("A")
+  for (i in theta) {
+    text(x + cos(i) * xo, y + sin(i) * yo, labels, col = bg, ...)
+  }
+  text(x, y, labels, col = col, ...)
 }